+++ CHANGELOG.md
@@ -76,6 +76,7 @@ Main (unreleased)
   This will load all River files in the directory as a single configuration;
   component names must be unique across all loaded files. (@rfratto, @hainenber)
 
+- Flow: Allow `prometheus.exporter.unix` to be specified multiple times and used in modules. (@mattdurham)
 
 ### Enhancements